Add and use modified wmctrl source code
In this commit only seeing that it works as a drop-in-replacement for
`wmctrl -l` for outputting the window list.

Unsure if Makefile will only work for me locally, or just fine in most
systems.

A more dynamic/automatic alternative to the `base_dir` variable would be
good to allow placing the project anywhere and only point to the main
script in polybar's config.ini `exec`.