Add sort options for windowlist
Polybar list automatic sorting, either:
* sort by horizontal position on screen
* alphabetically
* none (oldest first - newest last)

This is first iteration and seems good.

Issue with position sort:

What I really want is a way to "spy" the client list for window
position/geometry changes, so that the list could update right when a
window is moved - atm it can only update sorting order when changing
focus to different window and can be confusing to use if you don't know that.

Thanks to joooooooooooooooooooooooooooooooooooosh for the sorting method
in Bash:

https://github.com/joooooooooooooooooooooooooooooooooooosh/polywins/commit/991fa5e929156c680b3442a15f4d3c5fcfe43526

In case of same X-axis value, the window lower on Y-axis is considered further to the
right.

x and y refer to the point in the middle of a window - in vanilla
`wmctrl -lG` it is actually the top left corner, I don't like that for
sorting so I changed it.