Fix sorting on window position change (tentative)
A quick and dirty proof of concept where `xprop -root -spy` is replaced
by a similar loop, but instead of listening to PropertyNotify, it
listens to ConfigureNotify, so that the list is correctly sorted by
window position in any situation.

The problem before was that you could move or resize a window, but the
list would not be updated until window focus changes, or a window is
created/removed.

Note: Listening to ConfigureNotify alone would be enough as it also
sends when stacking order changes. However does this work on tiling
window managers? What if someone's WM doesn't raise window on focus
change? etc. Decided to be safe and keep PropertyNotify too.