Add user configurability
Finally bringing back (and improving on) the great customizability of
polywins.

Makes this module usable to other people without having to edit C code
and recompiling.

Also refactored `generate_window_list` so that it only takes windows
that are in the current desktop, this way we no longer have to care
about the desktop property anywhere else in the code.