Add spaces within window name, rename option
I realized it's significant whether the spaces are padded into the
string (window_name) or outside of it, when background colors will be
added. Checked whether polywins' `add_spaces` option has it so the
spaces have background color, or not. In polywins, the spaces have
background color too.

So to be equivalent to that, the spaces can be thought of as window name
padding to which bg color is applied. If spaces without bg color are
desired, they can be added to separator_string, for example:

separator_string = "  ·  "