Take config from `env-CONFIGPATH`, or default location
BREAKING CHANGES:

`windowlist` will no longer find `config.toml` in the root of the repo,
unless `env-CONFIGPATH` points to that.

To use a custom path for the TOML config file, set this in Polybar's
`config.ini`:

```
[module/windowlist]
env-CONFIGPATH = ~/any/path/windowlist.toml
```

If `env-CONFIGPATH` is not set, `windowlist` attempts to use either:
- `$XDG_CONFIG_HOME/polybar/windowlist.toml`
- `~/.config/polybar/windowlist.toml`

And if there's no file in this "default" path, `windowlist` falls back
to the default values of all options when running.